Repositories, queries, etc.

+++ src/main/scala/com/shin/db/persistence/AnimeRepository.scala
@@ -8,9 +8,9 @@ trait AnimeRepository[F[_]] {
 
   def insert(anime: Anime): F[Option[Anime]]
 
-  def get(id: Int): F[Option[Anime]]
+  def get(id: Long): F[Option[Anime]]
 
-  def delete(id: Int): F[Int]
+  def delete(id: Long): F[Int]
 
   def update(anime: Anime): F[Option[Anime]]
