+++ src/main/scala/com/shin/db/persistence/MovieRepository.scala
@@ -8,9 +8,9 @@ trait MovieRepository[F[_]] {
 
   def insert(movie: Movie): F[Option[Movie]]
 
-  def get(id: Int): F[Option[Movie]]
+  def get(id: Long): F[Option[Movie]]
 
-  def delete(id: Int): F[Int]
+  def delete(id: Long): F[Int]
 
   def update(movie: Movie): F[Option[Movie]]
